Create Gameboard() IIFE object containing game board and reset method
Created Gameboard() factory function wrapped inside an IIFE since only
single instance of it is needed (there is only one board in tic-tac-toe
after all).

Created an array representing the game board of length 9 as there are 9
boxes in tic-tac-toe and filled them with "empty" which represents an
empty slot for the player to fill.

resetBoard() method will serve to reset the board slots back to "empty"
when a game ends.