Create methods to check vacant board space and player move validation
checkSpaceVacancy() method loops through all the board slots and
returns true even if one slot is "empty" else returns false. This
is crucial for checking whether the game is over, especially in case
of Draws.

validateMove() method takes in the argument 'position' representing
the index number of the slot that the player wants to fill. The method
checks whether the slot is "empty" and returns boolean value.