Create method to update board with player input and checking win match
The updateBoard() method takes in the player input (X or O) and the
position representing the board slot to be filled and makes respective
changes to the game board by filling that particluar slot with input
provided.

The checkMatch() method checks for potential winning match across all
the rows, columns, and diagonals, and returns a boolean value.