Create Player factory function to store player data in object
The factory function takes in name and choice inputs where choice
refers to either 'X' or 'O'.

The playerInput() method takes input from player and returns it after
validation.

The playerMove() method is used to fill in the game board slot at
the index number of player's choice (by calling playerInput().)

The factory function returns an object containing the name of player,
their choice, and the playerMove() method.