Create IIFE for handling gameplay; make game console playable
The gameController is a factory function wrapped inside an IIFE.
Its purpose is to handle the game.

The playGame() method loops through both the players' turns until
either the board is filled or if one of the players have won.
It returns the result of the match.

Currently, the game has been made console playable with dummy
predefined players. The input is asked in the form of index numbers
via prompt().