Create html structure for main section
The game board is made up of buttons as they are meant to be clicked
by the players.

The buttons have been given an id starting with numbers preceeded
by 'b' as ids starting with number are invalid and not selectable.

The number in each button id corresponds to the location of that slot
in the gameboard array in script.