Finish content placement and minimal styling for main section
The score table has been set up using Grid layout as the data is
minimal and table layout would be better suited for when data is
more in quantity.

The divs with class '.commentary' and '.slash' have been hidden
and positioned absolutely as they might be used for showing comment
on the board itself upon game end and related stuff.

The board has been given a max-width of 100vw for display in mobile
devices.
The aspect ratio has been set to 1/1 to keep it in square shape.

Created a '--board-border' variable to serve as an input to the
border property of all the buttons of the game board since its value
is tentative and will be changed when styling starts.