Setup sidebar toggle button
Added a sidebar button to show and hide the sidebar with player
form.

This will be useful for narrow devices with lack of space so that
when the user is done entering the form, the sidebar is closed
when the game starts, giving more free space to the gameboard for
a more immersive experience.