Add functionality to handle player names in factory functions
The ScreenController factory function is back to being an IIFE as
only one instance of it was needed and it seemed better to just
implicitly name variables and manipulate them using methods that
are returned as part of object.

The same has been done in case of player name variables. The methods
to set and get player names are returned from ScreenController and
used as needed.