Add functionality to update score on DOM
Created two separate methods for updating score on DOM: one for
player X and another for player O.
This was done because the method was to be called inside the
GameController where the player objected created actually exists
inside which the score of the player is stored.

Therefore, in order to update score on DOM, we need separate
methods for each player.