Make name and choice variables of Player factory private
Instead of returning the name and choice variables themselves as
part of object, methods to get and set them have been created which
are to be returned instead of variables.

The parameter of setChoice() will default to being the same as
the existing choice in case no argument is provided.
This will be useful in case we switch players and want them to retain
their choice.