Create Player object inside ScreenController instead of GameController
The player object data is needed inside the ScreenController since
all the score update and gameboard update is to be done inside this
IIFE.

The method to get player name has been removed from ScreenController
since it is redundant as the Player object already provides a method
for the same.

The player object created inside the ScreenController is accessed
inside GameController IIFE by being passed as argument with the help
of getPlayerX() and getPlayerO() methods of ScreenController.