Refactor algorithm for reading in user input from DOM
Initially, the playerInput was an IIFE with an event listener
inside it. However, it only returned undefined.

Therefore, the event listener is now placed outside playerInput
and the id of the button clicked is stored in the variable boardSlot.
The playerInput is no longer an IIFE but a method which returns the
boardSlot value.