REDEFINE GameController, ScreenController, & Player functionalities
At some point, the functionalities of the GameController,
ScreenController, and Player factories became vague as they performed
multiple tasks including those which are not meant for them.

What is needed is a clear distinction and set of duties to be
performed especially and exclusively by each of them.

ScreenController will no longer be involved in an function that is
related to running the game. ScreenController will only be limited
to DOM Manipulation.

GameController will only be involved in handling the game itself.
Unfortunately, before this, ScreenController was seen to be taking
most of the functionalities that should have belonged to the
GameController.

Player factory function will only deal with the player data. It
should not be concerned with player moves. That is to be taken care
of by the GameController.
As a result, playerMove() method has been removed from Player factory.

Also, the Player object will no longer be instantiated within the
ScreenController; instead it is to be instantiated within the
GameController.