Add reset feature to GameController
Reset will be required in three situations:
1) Clicking the 'Start Game' button after filling Player form.
2) Clicking the 'RESET' button below the game board.
3) At the end of one round of game.

The first two options will require full reset (both scoreboard
and gameboard) while latter will only require a gameboard reset.

Therefore, the resetGameBoard() and reset() method of GameController
are returned and the former is called at the end of a game round
while the latter is called in situations (1) and (2).