Adjust player move switch algorithm
Initially, the player move was switched from within the event
listener itself because of which invalid clicks also led to the
switch.

Therefore, a switch method has been created which is called from
within the player move methods of GameController. This makes sure
that invalid clicks do not lead to pressing of the switch as well.