Fix post move checks feature to prevent final move WIN rendered as DRAW
Initialle the post move check methods first checked for a Win and then
for board space vacancy with both the tasks being in separate if
statements.
However, in case of final-move wins, both the if statements got
executed as the move was a win and also led to filling up of board
space vacancy.

Therefore, the issue has been fixed by placing the space vacancy check
in an 'else if' statement instead of if statement as either one of the
events take place.