Fix score-board position on top as user scrolls
In case the game-board is tall (especially in case of large devices),
the score-board went outside the view upon scrolling-down the page.

Thus, the user had to scroll-up to check score and scroll back down
to play the game which made things trickier.

Therefore, positioned the score-board stickily with main as relative
which ensures that score-board is always on display and user doesn't
have to scroll back up to view it.