Slightly thicken board border; style result display container
The board border seemed too thin as compared to the font thickness
of X and O. Therefore, it has been thickened by a small margin to
reduce the ratio.

The previous styling given to result display container was temporary.
A backdrop blur has been applied along with a dark background color
(in addition to transparency of parent container) for a stylish and
conspicuous view.

The Grid display of the text content divs of the result display has
been removed and a single Grid display has been applied to result
display and the items have been centered using place-items. This
saved some lines of code.

Padding block has been applied to give more space to text content
and gap has been applied to split them apart a bit.