Add feature to show a strikethrough on the winning triplet
The skeleton of the strikethrough was created in HTML & CSS using
<hr> tags, Grids, transform and various other properties.
The elements were set as hidden by default. All that was needed to
be done was to make specific <hr> visible depending upon the
position of the winning triplet.

The strikeWin() method of ScreenController has been created to take
in the 'position' (which is the specific id of <hr> element) of the
winning triplet and makes that particular <hr> visible.
The input is provided by the checkMatch() method of Gameboard which
has been updated to call strike() upon each winning condition with
the 'position' that corresponds to it.

The resetStrikeVisibility() has been created to turn visibility back
to 'hidden' when the result display is clicked.